pyproject.toml
[project]
name = "ozon-price-check"
version = "0.1.0"
description = "OZON product price checker and profit calculator TUI"
readme = "README.md"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "keyring>=25.6.0",
    "platformdirs>=4.3.8",
    "pydantic-settings>=2.10.1",
    "textual>=5.0.1",
]

[project.scripts]
ozon-price = "ozon_price_check.main:main"
price-check = "ozon_price_check.main:main"
